Refactor to pull out SandboxWrapper into sandbox.jsx
